requirements.txt
langchain-core
langchain-community
crewai[tools]
langgraph
langchain-groq
python-dotenv
litellm[proxy]
streamlit